Fix corecel errors/warnings on windows
Remove isatty for windows:
```
E:\scale\external\celeritas\src\corecel\io\ColorUtils.cc(13): fatal
error C1083: Cannot open include file: 'unistd.h': No such file or directory
```

String view iterator is not a const char*:
```
E:\scale\external\celeritas\src\corecel\io\StringUtils.cc(66): note:
'std::basic_string_view<char,std::char_traits<char>>::
basic_string_view(const char *const ,const
std::basic_string_view<char,std::char_traits<char>>::size_type)
noexcept': canno
t convert argument 1 from 'std::_String_view_iterator<_Traits>' to
'const char *const '
        with
    [
        _Traits=std::char_traits<char>
    ]
```

'environ' is a macro on windows (boo)
```
E:\scale\external\celeritas\src\corecel\sys\Environment.cc(31): error
C2267: '__p__environ': static functions with block scope are illegal
```

psapi must be included *after* windows.h
```
C:\Program Files (x86)\Windows
Kits\10\\include\10.0.22621.0\\um\psapi.h(132): error C2146: syntax error: missing ';' before identifier 'WINAPI'
```

ptrdiff_t is implicitly converted to long
```
E:\scale\external\celeritas\src\corecel\sys\ScopedMem.cc(125): warning
C4244: 'argument': conversion from 'ptrdiff_t' to 'long', possible loss
of dat
a
```

+++ src/corecel/sys/MemRegistry.hh
@@ -20,12 +20,12 @@ namespace celeritas
 //! SI prefix for multiples of 1024
 struct Kibi
 {
-    static CELER_CONSTEXPR_FUNCTION int value() { return 1024; }
+    static CELER_CONSTEXPR_FUNCTION long int value() { return 1024; }
     static char const* label() { return "kibi"; }
 };
 
 //! 1024 bytes
-using KibiBytes = Quantity<Kibi, int>;
+using KibiBytes = Quantity<Kibi, long int>;
 //! Ordered identifiers for memory allocation segments
 using MemUsageId = OpaqueId<struct MemUsageEntry>;
